Include schema property name metadata on required properties
For consistency with Python conventions, jschema-to-python maps property
names in the JSON schema (which are typically camelCase) to
underscore_separated attribute names in the generated Python classes.
The generated attributes include metadata that maps the Python property
names back to the original JSON schema property names. The to_json
method uses this metadata so that object models built with the generated
Python classes are correctly serialized to JSON that conforms to the
schema.

The bug here was that although optional properties contained this
metadata, required properties did not. This meant that if a required
JSON schema property consisted of more than one word (for example,
"fileChanges"), the Python object containing that property would be
serialized as "file_changes".

We have regenerated the SARIF OM classes with a new version of
jschema-to-python that fixes this bug.